deleted src folder; test_app.py moved to the core
