+++ insert_data.py
@@ -1,4 +1,4 @@
-from src.models import LapTime, Racer
+from models import LapTime, Racer
 
 
 def insert_report(racers: list[LapTime]) -> list[dict]:
